requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pydantic-settings>=2.0.0
pdf2image>=1.17.0
Pillow>=10.0.0
gradio_client>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
